+++ vitest.config.ts
@@ -1,0 +1,19 @@
+import path from "path";
+import react from "@vitejs/plugin-react";
+import tailwindcss from "@tailwindcss/vite";
+import { defineConfig } from "vitest/config";
+
+// https://vite.dev/config/
+export default defineConfig({
+  test: {
+    environment: "jsdom", // Correct environment for React Testing Library
+    // setupFiles: "./src/setupTests.ts", // Path to your setup file for jest-dom
+    globals: true, // This enables `expect` to work globally
+  },
+  plugins: [react(), tailwindcss()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
+});